+++ cmd/hidden_lake/service/internal/handler/network_request.go
@@ -109,7 +109,7 @@ func HandleNetworkRequestAPI(
 			}
 
 			pLogger.PushInfo(logBuilder.WithMessage(http_logger.CLogSuccess))
-			_ = api.Response(pW, http.StatusOK, resp)
+			_ = api.Response(pW, http.StatusOK, resp.ToString())
 			return
 		}
 	}